refactor(dashboard): move Highlights to Insights + fix <synthetic> filter
Two coupled changes addressing dashboard density and a pre-existing
advisor bug the user surfaced while checking the new panel:

1. Move Highlights from Dashboard to Insights.

   Dashboard had ~12 quantitative sections; adding Highlights on top
   pushed everything further down and mixed content-curation with
   aggregate charts. Insights is the natural home — it already pairs
   Trends (quantitative) with Recommendations (prescriptive); Highlights
   (descriptive content) completes the retrospective arc:

     Headline → Summary stats → Model info →
       [new] Highlights →
     Trends → Duration vs Quality → Recommendations

   The panel now respects the Insights tab's `days` selector (7 / 14 /
   30) rather than the hardcoded 7-day window it had on the Dashboard —
   so "show me my top 3 of the last month" works naturally. The endpoint
   and the 3 / min_quality=4 card count stay fixed; only the window
   tracks the user's selection.

   Dashboard.tsx: drop HighlightCard, helpers, state, fetch, and the
   section render. Net -~140 lines.

   Insights.tsx: add HighlightCard + helpers (same shape as before),
   fetch highlights alongside advisor + insights via Promise.all with
   a per-call catch so a highlights failure can't block the tab.

2. Fix <synthetic> filter in scoring/insights.py.

   `<synthetic>` is a parser fallback model name used when a session's
   real model isn't detectable. The export path at cli.py:479 already
   drops those sessions, but the advisor's by_model query did not —
   it grouped them together, and since they usually have cost=0, they
   trivially won "Most Efficient" (cost/sessions = 0) and could slip
   into "Highest Quality" on a lucky score average. Mirror the export
   filter: `AND model != '<synthetic>'`. One-line change; live check
   now shows real model names (gpt-5.4 / claude-opus-4-6) in both
   cards.

Tests: full suite 1003 green (no test covered the advisor's
by_model filter directly, so no test changed — worth a future followup
but not blocking). Frontend build clean. Endpoint verified live with
a running daemon.
