+++ clawjournal/web/frontend/src/views/Dashboard.tsx
@@ -1,6 +1,5 @@
 import { useState, useEffect, useCallback } from 'react';
-import { Link } from 'react-router-dom';
-import type { DashboardData, HighlightItem, HighlightsData, InsightsData } from '../types.ts';
+import type { DashboardData, InsightsData } from '../types.ts';
 import { api } from '../api.ts';
 import { LABELS } from '../components/BadgeChip.tsx';
 import { Spinner } from '../components/Spinner.tsx';
@@ -79,126 +78,6 @@ function Section({ title, subtitle, children }: { title: string; subtitle?: stri
       </div>
       {children}
     </div>
-  );
-}
-
-function shortDuration(seconds: number | null | undefined): string {
-  if (!seconds || seconds <= 0) return '—';
-  const hours = Math.floor(seconds / 3600);
-  const minutes = Math.floor((seconds % 3600) / 60);
-  if (hours && minutes) return `${hours}h ${minutes}m`;
-  if (hours) return `${hours}h`;
-  if (minutes) return `${minutes}m`;
-  return `${seconds}s`;
-}
-
-function sourceLabel(source: string | null | undefined): string {
-  if (!source) return '';
-  return source.charAt(0).toUpperCase() + source.slice(1);
-}
-
-function outcomeColor(outcome: string | null | undefined): string {
-  if (!outcome) return colors.gray400;
-  const resolved = ['resolved', 'shipped', 'tests_passed', 'completed', 'success'];
-  const bad = ['failed', 'abandoned', 'build_failed', 'tests_failed', 'errored'];
-  if (resolved.includes(outcome)) return colors.green500;
-  if (bad.includes(outcome)) return colors.red400;
-  return colors.yellow400;
-}
-
-function HighlightCard({ item }: { item: HighlightItem }) {
-  const scoreColor = item.ai_quality_score && item.ai_quality_score >= 5
-    ? colors.green500
-    : item.ai_quality_score && item.ai_quality_score >= 4
-    ? colors.blue400
-    : colors.gray400;
-
-  return (
-    <Link
-      to={`/session/${item.session_id}`}
-      style={{
-        display: 'block',
-        textDecoration: 'none',
-        color: 'inherit',
-        border: `1px solid ${colors.gray200}`,
-        borderRadius: 8,
-        padding: 12,
-        background: '#fff',
-        minHeight: 156,
-        flex: '1 1 260px',
-        maxWidth: 360,
-        transition: 'border-color 0.15s, box-shadow 0.15s',
-      }}
-      onMouseEnter={(e) => {
-        e.currentTarget.style.borderColor = colors.blue400;
-        e.currentTarget.style.boxShadow = '0 2px 6px rgba(0,0,0,0.06)';
-      }}
-      onMouseLeave={(e) => {
-        e.currentTarget.style.borderColor = colors.gray200;
-        e.currentTarget.style.boxShadow = 'none';
-      }}
-    >
-      <div style={{ display: 'flex', justifyContent: 'space-between', gap: 6, marginBottom: 6 }}>
-        <div style={{ display: 'flex', gap: 6, flexWrap: 'wrap' }}>
-          {item.ai_quality_score != null && (
-            <span style={{
-              fontSize: 11, fontWeight: 600, color: scoreColor,
-              background: `${scoreColor}18`, padding: '2px 7px', borderRadius: 10,
-            }}>
-              {item.ai_quality_score}/5
-            </span>
-          )}
-          {item.outcome && (
-            <span style={{
-              fontSize: 11, fontWeight: 500, color: outcomeColor(item.outcome),
-              background: `${outcomeColor(item.outcome)}15`, padding: '2px 7px', borderRadius: 10,
-            }}>
-              {labelFor(item.outcome)}
-            </span>
-          )}
-        </div>
-        {item.source && (
-          <span style={{
-            fontSize: 10, fontWeight: 600, color: colors.gray600,
-            textTransform: 'uppercase', letterSpacing: 0.3,
-          }}>
-            {sourceLabel(item.source)}
-          </span>
-        )}
-      </div>
-
-      <div style={{
-        fontSize: 13, fontWeight: 600, color: colors.gray900, marginBottom: 6,
-        display: '-webkit-box', WebkitLineClamp: 2, WebkitBoxOrient: 'vertical',
-        overflow: 'hidden', lineHeight: 1.3,
-      }}>
-        {item.title}
-      </div>
-
-      {item.summary_teaser && (
-        <div style={{
-          fontSize: 12, color: colors.gray700, marginBottom: 8, lineHeight: 1.45,
-          display: '-webkit-box', WebkitLineClamp: 3, WebkitBoxOrient: 'vertical',
-          overflow: 'hidden',
-        }}>
-          {item.summary_teaser}
-        </div>
-      )}
-
-      <div style={{
-        display: 'flex', justifyContent: 'space-between', alignItems: 'center',
-        gap: 6, fontSize: 11, color: colors.gray500,
-      }}>
-        <span style={{
-          overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap', flex: 1,
-        }} title={item.project ?? ''}>
-          {displayProject(item.project ?? '—')} · {shortDuration(item.duration_seconds)}
-        </span>
-        <span style={{ color: colors.gray500, fontStyle: 'italic', flexShrink: 0 }}>
-          {item.rationale}
-        </span>
-      </div>
-    </Link>
   );
 }
 
@@ -286,7 +165,6 @@ export function Dashboard() {
   const [data, setData] = useState<DashboardData | null>(null);
   const [triage, setTriage] = useState<TriageStats | null>(null);
   const [insights, setInsights] = useState<InsightsData | null>(null);
-  const [highlights, setHighlights] = useState<HighlightsData | null>(null);
   const [timeRange, setTimeRange] = useState<string>('1w');
   const [loading, setLoading] = useState(true);
 
@@ -307,13 +185,6 @@ export function Dashboard() {
         setInsights(await api.insights(dashParams));
       } catch {
         setInsights(null);
-      }
-      // Fetch highlights — always last 7 days regardless of timeRange, since
-      // the panel is curation-across-agents rather than a filtered view.
-      try {
-        setHighlights(await api.highlights({ days: 7, top: 3, min_quality: 4 }));
-      } catch {
-        setHighlights(null);
       }
     } catch (e) {
       toast(e instanceof Error ? e.message : 'Failed to load dashboard', 'error');
@@ -402,28 +273,6 @@ export function Dashboard() {
         ))}
       </div>
 
-      {/* Highlights — top 3 five-star sessions across different agents, last 7 days.
-          Independent of the timeRange selector by design: this is a curation
-          panel, not a filtered view. */}
-      {highlights && highlights.highlights.length > 0 && (
-        <Section
-          title="Highlights"
-          subtitle={`Top ${highlights.highlights.length} five-star sessions across agents, last ${highlights.window_days} days`}
-        >
-          <div style={{ display: 'flex', gap: 10, flexWrap: 'wrap' }}>
-            {highlights.highlights.map((item) => (
-              <HighlightCard key={item.session_id} item={item} />
-            ))}
-          </div>
-        </Section>
-      )}
-      {highlights && highlights.highlights.length === 0 && (
-        <Section title="Highlights" subtitle={`No 4+ star sessions in the last ${highlights.window_days} days`}>
-          <div style={{ fontSize: 12, color: colors.gray500, padding: 8 }}>
-            Run <code style={{ background: colors.gray100, padding: '1px 5px', borderRadius: 3 }}>clawjournal score</code> on recent sessions to populate this panel.
-          </div>
-        </Section>
-      )}
 
       {/* Outcomes + Agent Behavior — promoted for visibility */}
       <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 12, marginBottom: 12 }}>
